Cygwin: Generate dummy file descriptors
The consumer in LTTng-tools keeps a mapping of the file descriptors
between the sessiond and it's own file descriptors in a hash table.
The fd number is the key used for lookups in the hash table.
Since we can't send the fds to the sessiond on Cygwin, the consumer
fds mapping mechanism is broken.

To fix this, we generate a dummy file descriptor upon reception
of the file paths in the sessiond.