Cygwin: Fix multiple loaded instances of liblttng-ust-tracepoint due to dlopen
Multiple instances of the liblttng-ust-tracepoint were present in memory and
caused duplication and incoherence of various data structures (such as the
tracepoints hashtable). This is due to the dlopen(3) of liblttng-ust-tracepoint
in applications that were already linked against liblttng-ust-tracepoint.

This commit introduce a new restriction on application tracing:
the applications must be linked against liblttng-ust-tracepoint.